mosquitto-docker/python-app/requirements.txt
autopep8
beautifultable
click<8.0,>=5.1
CounterFit
counterfit-connection
counterfit-shims-grove
dnspython==2.2.1
Flask==1.1.4
gitpython
ipykernel
ipython
markupsafe==2.0.1
matplotlib
mysql-connector-python
numpy<2.0.0,>=1.23.5
paho-mqtt
pandas
pip
scikit-learn
setuptools
wheel
yapf